% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{cdec_csv_parse}
\alias{cdec_csv_parse}
\title{Parse CDEC CSV}
\usage{
cdec_csv_parse(filename)
}
\arguments{
\item{filename}{Path to a downloaded CDEC file in the format output by the CDEC \code{CSVDataServlet} service, such as retrieved by the \code{retrieve_cdec_csv()} function. May be compressed (\code{.csv.gz}) or uncompressed (\code{.csv}).}
}
\value{
A \code{tbl_df} data frame with cleaned tabular data
}
\description{
Parse CDEC CSV
}
