% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{cdec_csv_retrieve}
\alias{cdec_csv_retrieve}
\title{Retrieve CDEC CSV}
\usage{
cdec_csv_retrieve(
  sta = character(),
  sen = character(),
  dur = character(),
  start_date = lubridate::ymd("1900-10-01"),
  end_date = lubridate::ymd("2023-09-30"),
  dir = "temp"
)
}
\arguments{
\item{sta}{3-character alphanumeric code for CDEC station}

\item{sen}{Numeric code for CDEC sensor}

\item{dur}{1-character code for duration type: \code{D}=daily, \code{H}=hourly, \code{E}=event (15 minute)}

\item{start_date}{Start date for query expressed as R date object e.g. lubridate::ymd("1900-10-01")}

\item{end_date}{End date for query expressed as R date object e.g. lubridate::ymd("2023-09-30")}

\item{dir}{Directory to store outputs in}
}
\value{
The path to the downloaded CDEC file in \code{csv.gz} (compressed CSV) format
}
\description{
Retrieve CDEC CSV
}
