% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{const_proj_crs}
\alias{const_proj_crs}
\title{Project CRS}
\usage{
const_proj_crs()
}
\description{
Project CRS
}
\examples{
sf::st_sfc(crs="+proj=latlong +datum=WGS84") |>
  sf::st_transform(const_proj_crs())
sf::st_crs(const_proj_crs())
}
