% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cummean.na.rm}
\alias{cummean.na.rm}
\title{Cumulative Mean ignoring NA values}
\usage{
cummean.na.rm(x)
}
\arguments{
\item{x}{An integer or numeric vector}
}
\value{
A vector the same length as \code{x}
}
\description{
Cumulative Mean ignoring NA values
}
\examples{
v <- c(1, 2, 3, NA, NA, 1, 2, 6, NA)
cummean.na.rm(v)

}
\keyword{internal}
