% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv_mainstems}
\alias{cv_mainstems}
\title{CVPIA Mainstems (geometry)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3366 rows and 5 columns.
}
\usage{
cv_mainstems
}
\description{
CVPIA Mainstems (geometry)
}
\examples{
cv_mainstems |> ggplot2::ggplot() + ggplot2::geom_sf()
}
\keyword{datasets}
