% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv_watersheds}
\alias{cv_watersheds}
\title{CVPIA Watersheds (geometry)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 85 rows and 8 columns.
}
\usage{
cv_watersheds
}
\description{
CVPIA Watersheds (geometry)
}
\examples{
cv_watersheds |> ggplot2::ggplot() + ggplot2::geom_sf()
}
\keyword{datasets}
