% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cv_watersheds_flow_xw}
\alias{cv_watersheds_flow_xw}
\title{Flow Area Crosswalk for CV Watersheds}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 67145 rows and 8 columns.
}
\usage{
cv_watersheds_flow_xw
}
\description{
Flow Area Crosswalk for CV Watersheds
}
\keyword{datasets}
