% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{drive_file_by_id}
\alias{drive_file_by_id}
\title{Download Data from Google Drive}
\usage{
drive_file_by_id(
  id = character(),
  dir = here::here("data-raw", "temp"),
  vsizip = F
)
}
\arguments{
\item{id}{The character ID from a Google Drive file. Found in the middle of the share link: \verb{https://drive.google.com/file/d/XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/view}}

\item{dir}{Path to the local directory in which to download the file}

\item{vsizip}{Whether to append the "vsizip" parameter which allows GEOS access within ZIPs without extracting first}
}
\value{
The path to the downloaded file
}
\description{
Download Data from Google Drive
}
