% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration_apply_dhsi_to_fsa_curve}
\alias{duration_apply_dhsi_to_fsa_curve}
\title{Apply Duration HSI to Flow-to-Suitable-Area Curve}
\usage{
duration_apply_dhsi_to_fsa_curve(
  fsa,
  drc,
  fsa_q = q,
  fsa_wua = wua,
  drc_q = q,
  drc_dhsi = dhsi
)
}
\arguments{
\item{fsa}{A \code{data.frame} or \code{tbl_df} containing a flow-to-suitable-area curve}

\item{drc}{A \code{data.frame} or \code{tbl_df} containing a flow-to-duration-HSI curve}

\item{fsa_q}{The unquoted name of the flow column in \code{fsa} (defaults to \code{q})}

\item{fsa_wua}{The unquoted name of the suitable area column in \code{fsa} (defaults to \code{wua})}

\item{drc_q}{The unquoted name of the flow column in \code{drc} (defaults to \code{q})}

\item{drc_dhsi}{The unquoted name of the duration suitability factor column in \code{drc} (defaults to \code{dhsi})}
}
\value{
A \code{tbl_df} data frame with one row per flow (\code{q}) containing columns for the original suitable area (\code{wua}) and the new duration-weighted suitable area (\code{durwua}) along with interim calculation columns.
}
\description{
Apply Duration HSI to Flow-to-Suitable-Area Curve
}
