% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration_calc_days_inundated}
\alias{duration_calc_days_inundated}
\title{Calculate Days Inundated}
\usage{
duration_calc_days_inundated(q_gauge, q_crit, stat = "max")
}
\arguments{
\item{q_gauge}{A vector of flow values, typically a time series}

\item{q_crit}{A numeric value against which to calculate the exceedance}

\item{stat}{Either "max" to calculate maximum days inundated (default) or "sum" to calculate total days inundated}
}
\value{
A single numeric value for the specified statistic.
}
\description{
Calculate Days Inundated
}
