% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration_hsi_rearing_vf}
\alias{duration_hsi_rearing_vf}
\title{Rearing Duration HSI for Valley Foothill}
\usage{
duration_hsi_rearing_vf(x, baseflow_threshold_days = Inf)
}
\arguments{
\item{x}{A vector of values of number of days}

\item{baseflow_threshold_days}{An optional threshold number of days defining baseflow}
}
\value{
A vector of values of duration HSI
}
\description{
Rearing Duration HSI for Valley Foothill
}
