% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration_hsi_spawning}
\alias{duration_hsi_spawning}
\title{Spawning Duration HSI}
\usage{
duration_hsi_spawning(x)
}
\arguments{
\item{x}{A vector of values of number of days}
}
\value{
A vector of values of duration HSI
}
\description{
Spawning Duration HSI
}
