% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{flow_threshold_remove_baseflow}
\alias{flow_threshold_remove_baseflow}
\title{Remove Baseflow via Flow Threshold (Single)}
\usage{
flow_threshold_remove_baseflow(
  data,
  flow_threshold,
  .flow_var = flow_cfs,
  .habitat_var = wua_per_lf
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl_df} containing a flow-to-suitable-area curve}

\item{flow_threshold}{A numeric value for a flow threshold defining baseflow.}

\item{.flow_var}{Unquoted column name for the flow in cfs, defaults to \code{flow_cfs}}

\item{.habitat_var}{Unquoted column name for the suitable area, defaults to \code{wua_per_lf}}
}
\value{
A \code{tbl_df} data frame with one row per flow containing two columns: the flow and the modified suitable habitat area with baseflow removed.
}
\description{
Remove Baseflow via Flow Threshold (Single)
}
