% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{flow_threshold_split}
\alias{flow_threshold_split}
\title{Split Flow-to-Suitable-Area Curve at Flow Threshold (Helper Function)}
\usage{
flow_threshold_split(suitable_area, flow, flow_threshold)
}
\arguments{
\item{suitable_area}{A numeric vector of suitable area values, e.g. in ft2 per lf}

\item{flow}{A numeric vector of flow values in cfs, same length as \code{suitable_area}}

\item{flow_threshold}{A numeric value for the instream-floodplain flow threshold in cfs}
}
\value{
A named list including the original \code{suitable_area} and \code{flow} vectors, an identifier of instream versus floodplain (\code{above_threshold}), and suitable area values for instream and floodplain (\code{habitat_below} and \code{habitat_above})
}
\description{
Split Flow-to-Suitable-Area Curve at Flow Threshold (Helper Function)
}
