% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{flow_threshold_split_apply}
\alias{flow_threshold_split_apply}
\title{Split Flow-to-Suitable-Area Curve at Flow Threshold}
\usage{
flow_threshold_split_apply(
  data,
  flow_threshold = NULL,
  .flow_threshold_var = NULL,
  .id_var = comid,
  .flow_var = flow_cfs,
  .habitat_var = wua_per_lf,
  .group_var = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl_df} containing flow-to-suitable area curves (one row per flow per site, columns for flow and suitable area)}

\item{flow_threshold}{A numeric value for a flow threshold defining the instream-floodplain split}

\item{.flow_threshold_var, }{An unquoted column name for the flow threshold column in the data frame. Ignored if \code{flow_threshold} is provided.}

\item{.id_var}{Unquoted column name for the variable id, defaults to \code{comid}}

\item{.flow_var}{Unquoted column name for the flow in cfs, defaults to \code{flow_cfs}}

\item{.habitat_var}{Unquoted column name for the suitable area, defaults to \code{wua_per_lf}}

\item{.group_var}{Unquoted column name for an optional grouping variable}
}
\value{
A new version of the \code{data.frame} or \code{tbl_df} containing additional columns for the instream and floodplain (below or above threshold) split
}
\description{
Split Flow-to-Suitable-Area Curve at Flow Threshold
}
