% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowline_attr}
\alias{flowline_attr}
\title{NHD Flowline (ComID) Attributes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 178868 rows and 162 columns.
}
\usage{
flowline_attr
}
\description{
NHD Flowline (ComID) Attributes
}
\examples{
flowline_attr |> pillar::glimpse()
}
\keyword{datasets}
