% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowline_geom}
\alias{flowline_geom}
\title{NHD Flowline (ComID) Geometries}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 178868 rows and 2 columns.
}
\usage{
flowline_geom
}
\description{
NHD Flowline (ComID) Geometries
}
\examples{
flowline_geom |> pillar::glimpse()
}
\keyword{datasets}
