% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowline_geom_proj}
\alias{flowline_geom_proj}
\title{NHD Flowline (ComID) Geometries - Projected}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 178868 rows and 2 columns.
}
\usage{
flowline_geom_proj
}
\description{
NHD Flowline (ComID) Geometries - Projected
}
\keyword{datasets}
