% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habistat-package.R
\docType{package}
\name{habistat-package}
\alias{habistat}
\alias{habistat-package}
\title{habistat: HabiStat}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Description here
}
\author{
\strong{Maintainer}: Skyler Lewis \email{slewis@flowwest.com}

Authors:
\itemize{
  \item Maddee Rubenson \email{mrubenson@flowwest.com} [project director]
}

}
\keyword{internal}
