% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hex_color_blend}
\alias{hex_color_blend}
\title{Hex Color Blend}
\usage{
hex_color_blend(x, y)
}
\arguments{
\item{x}{A character string describing a color in the form \code{"#000000"}}

\item{y}{A character string describing a color in the form \code{"#000000"}}
}
\value{
A character string describing a color in the form \code{"#000000"}
}
\description{
Hex Color Blend
}
\examples{
hex_color_blend("#55ad89", "#c3bc3f")

}
\keyword{internal}
