% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hqt}
\alias{hqt}
\title{HQT Valley Lowland Polygon}
\format{
An object of class \code{sfc_GEOMETRY} (inherits from \code{sfc}) of length 1020.
}
\usage{
hqt
}
\description{
HQT Valley Lowland Polygon
}
\keyword{datasets}
