% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{linterp}
\alias{linterp}
\title{Simple Linear Interpolation}
\usage{
linterp(x, x1, x2, y1, y2)
}
\arguments{
\item{x}{The index at which to predict \code{y}}

\item{x1}{An index \code{x1 <= x} with known value \code{y1}}

\item{x2}{An index \code{x1 >= x} with known value \code{y2}}

\item{y1}{The value at index \code{x1}}

\item{y2}{The value at index \code{x2}}
}
\value{
The predicted value \code{y} at index \code{x}
}
\description{
Simple Linear Interpolation
}
