% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{oom_range}
\alias{oom_range}
\title{Order of Magnitude Range}
\usage{
oom_range(from, to)
}
\arguments{
\item{from}{An integer defining the start of the sequence}

\item{to}{An integer defining the end of the sequence}
}
\value{
An integer vector sequence
}
\description{
Order of Magnitude Range
}
\examples{
oom_range(9, 47000)

}
\keyword{internal}
