% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perpendicular_transect}
\alias{perpendicular_transect}
\title{Perpendicular Transect (helper function for perpendicular_transects)}
\usage{
perpendicular_transect(pair = sf::st_multipoint(), length = numeric())
}
\arguments{
\item{pair}{An \code{sf} multipoint containing two points along a stream or profile line}

\item{length}{The length of transect to create}
}
\value{
An \code{sf} linestring transect, halfway between the two input points in \code{pair}, angled perpendicular to their alignment, measuring length \code{length}.
}
\description{
Perpendicular Transect (helper function for perpendicular_transects)
}
