% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perpendicular_transects}
\alias{perpendicular_transects}
\title{Create Perpendicular Transects from Line}
\usage{
perpendicular_transects(ls = sf::st_linestring(), length = numeric())
}
\arguments{
\item{ls}{An \code{sfc} linestring geometry collection defining a stream or profile line along which to generate transects.}

\item{length}{The length of each transect to create, which typically should be wide enough to span the streamway or valley in question.}
}
\value{
An \code{sfc} linestring geometry collection of perpendicular transect lines, one for each pair of subsequent points in \code{ls}
}
\description{
Create Perpendicular Transects from Line
}
