% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_dvhsi_hqt}
\alias{raster_dvhsi_hqt}
\title{Raster Habitat Suitability Index (HSI) function: HQT}
\usage{
raster_dvhsi_hqt(d, v)
}
\arguments{
\item{d}{a \code{terra} raster layer for depth in feet}

\item{v}{a \code{terra} raster layer for velocity in feet per second}
}
\value{
A \code{terra} raster layer of habitat suitability index values between 0 and 1.
}
\description{
Raster Habitat Suitability Index (HSI) function: HQT
}
