% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_dvhsi_spawning}
\alias{raster_dvhsi_spawning}
\title{Raster Habitat Suitability Index (HSI) function: Fall-Run Chinook Spawning}
\usage{
raster_dvhsi_spawning(d, v, run = "max", strict = T)
}
\arguments{
\item{d}{a \code{terra} raster layer for depth in feet}

\item{v}{a \code{terra} raster layer for velocity in feet per second}
}
\value{
A \code{terra} raster layer of habitat suitability index values between 0 and 1.
}
\description{
Raster Habitat Suitability Index (HSI) function: Fall-Run Chinook Spawning
}
