% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_first_inundating_flow}
\alias{raster_first_inundating_flow}
\title{Calculate First Inundating Flow}
\usage{
raster_first_inundating_flow(rasters)
}
\arguments{
\item{rasters}{A named list or \code{SpatRasterDataset} of \code{depth} and \code{velocity} raster stacks (in \code{terra} format) containing model results for each flow, as output by \code{raster_prep_grid()}. Only depth is strictly necessary as velocity is not used in this function. Layer names should be the flow in cfs.}
}
\value{
A \code{terra} \code{SpatRaster} in which each grid cell is identified by the first modeled flow where depth exceeded 0 ft.
}
\description{
Calculate First Inundating Flow
}
