% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_prep_grid}
\alias{raster_prep_grid}
\title{Prep depth and velocity raster stacks}
\usage{
raster_prep_grid(filenames)
}
\arguments{
\item{filenames}{A data frame or tibble containing the file path to each depth and velocity raster at each flow. There should be a numeric column called \code{flow_cfs}, and two character columns called \code{depth} and \code{velocity} each containing valid file paths to rasters (e.g. \code{.tif} or \code{.vrt} format).}
}
\value{
A named list of two \code{terra} \code{SpatRaster} raster stacks (\code{depth} and \code{velocity}), each containing depth or velocity grids at each modeled flow.
}
\description{
Prep depth and velocity raster stacks
}
