% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_summarize_hsi}
\alias{raster_summarize_hsi}
\title{Calculate HSI based on HEC-RAS 2D rasters}
\usage{
raster_summarize_hsi(
  rasters,
  group_polygons = NULL,
  mask_polygons = NULL,
  mask_raster = NULL,
  hsi_func = raster_dvhsi_hqt,
  .group_var = comid,
  parallel = FALSE
)
}
\arguments{
\item{rasters}{A named list or \code{SpatRasterDataset} of \code{depth} and \code{velocity} raster stacks (in \code{terra} format) containing model results for each flow, as output by \code{raster_prep_grid()}}

\item{group_polygons}{An \code{sf} polygon feature collection with one polygon for each analysis reach. These may have been delineated previously by splitting the hydraulic model domain into subsections.}

\item{mask_polygons}{Optional: An \code{sf} polygon feature or feature collection containing a mask to be eliminated from the study area; for example, a baseflow channel. Only works if group polygons are also supplied.}

\item{mask_raster}{Optional: A \code{terra} raster of 1 or 0 to be multipled for the purpose of masking out a baseflow channel.}

\item{hsi_func}{A \code{terra::lapp}-compatible function that takes depth (ft) and velocity (ft/s) raster layers and returns habitat suitability index values between 0 and 1. Function arguments should be \code{d} and \code{v} in US units. Defaults to the \code{raster_dvhsi_hqt} function in this package.}

\item{.group_var}{The unquoted name of the attribute used to identify groups in the \code{group_polygons} layer. Defaults to \code{comid}}

\item{parallel}{Toggles parallel processing. Currently under development, keep \code{FALSE} for now.}
}
\value{
A \code{tbl_df} data frame with one row per group (\code{comid}) per flow (\code{flow_cfs}), containing the total inundated area (\code{area_tot}) in square feet, the suitable habitat area (\code{area_wua}) in square feet, and the ratio of suitable to total area (\code{area_pct}).
}
\description{
Calculate HSI based on HEC-RAS 2D rasters
}
