% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{semiIHS00_inv}
\alias{semiIHS00_inv}
\title{Semi Inverse Hyperbolic Sine Transform * 100 (Inverse)}
\usage{
semiIHS00_inv(y)
}
\description{
Semi Inverse Hyperbolic Sine Transform * 100 (Inverse)
}
\keyword{internal}
