% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{seq_log10}
\alias{seq_log10}
\title{Logarithmic Sequence}
\usage{
seq_log10(from, to, by = 0.1, snap)
}
\arguments{
\item{from}{An integer defining the start of the sequence}

\item{to}{An integer defining the end of the sequence}

\item{by}{A number defining the exponent of 10 to increment by. Defaults to 0.1.}

\item{snap}{An optional integer defining a midpoint of the sequence that should be preserved exactly in the range. If not defined, then \code{from} will be preserved exactly.}
}
\value{
An integer vector sequence
}
\description{
Logarithmic Sequence
}
\examples{
seq_log10(from = 50, to = 15000, by = 0.1, snap = 100)

}
\keyword{internal}
