% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{st_coalesce}
\alias{st_coalesce}
\title{Spatial coalesce}
\usage{
st_coalesce(x, y, radius = 0)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{y}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{radius}{(optional) numeric or object of class \code{units}; radius for the morphological opening and closing operations that are applied if \code{radius != 0}. In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{TRUE}, a numeric \code{radius} is taken as distance in meters and a \code{units} object in \code{radius} is converted to meters. In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{FALSE}, a numeric \code{radius} is taken as degrees, and a \code{units} object in \code{radius} is converted to \code{arc_degree} (and warnings are issued). In case \code{x} does not have geodetic coordinates (projected) then numeric \code{radius} is assumed to have the units of the coordinates, and a \code{units} \code{radius} is converted to those if \code{st_crs(x)} is not \code{NA}.}
}
\value{
object of class \code{sf}, \code{sfc} or \code{sfg}, a topological mosaic that includes \code{x} geometries within the footprint of \code{x} and \code{y} geometries outside of the footprint of \code{x}. Can be applied to mosaic a list of more than two layers by using \code{purrr::reduce(list(x, y, z, ...), st_coalesce)}.
}
\description{
Spatial coalesce
}
