% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{st_identity}
\alias{st_identity}
\title{Spatial identity}
\usage{
st_identity(x, y, radius = 0)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{y}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{radius}{(optional) numeric or object of class \code{units}; radius for the morphological opening and closing operations that are applied if \code{radius != 0}. In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{TRUE}, a numeric \code{radius} is taken as distance in meters and a \code{units} object in \code{radius} is converted to meters. In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{FALSE}, a numeric \code{radius} is taken as degrees, and a \code{units} object in \code{radius} is converted to \code{arc_degree} (and warnings are issued). In case \code{x} does not have geodetic coordinates (projected) then numeric \code{radius} is assumed to have the units of the coordinates, and a \code{units} \code{radius} is converted to those if \code{st_crs(x)} is not \code{NA}.}
}
\value{
object of class \code{sf}, \code{sfc} or \code{sfg}, a "left join" style union operation that includes the full footprint of \code{x} and those portions of \code{y} that overlap \code{x}
}
\description{
Spatial identity
}
