% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{st_morph_close}
\alias{st_morph_close}
\title{Morphological closing for vector polygon geometry (eliminate scraps)}
\usage{
st_morph_close(x, radius = 0.01)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}

\item{radius}{numeric or object of class \code{units}; radius for the morphological closing operation (erosion followed by dilation). In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{TRUE}, a numeric \code{radius} is taken as distance in meters and a \code{units} object in \code{radius} is converted to meters. In case \code{x} has geodetic coordinates (lon/lat) and \code{sf_use_s2()} is \code{FALSE}, a numeric \code{radius} is taken as degrees, and a \code{units} object in \code{radius} is converted to \code{arc_degree} (and warnings are issued). In case \code{x} does not have geodetic coordinates (projected) then numeric \code{radius} is assumed to have the units of the coordinates, and a \code{units} \code{radius} is converted to those if \code{st_crs(x)} is not \code{NA}.}
}
\value{
object of class \code{sf}, \code{sfc} or \code{sfg}
}
\description{
Morphological closing for vector polygon geometry (eliminate scraps)
}
