% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamgage_attr}
\alias{streamgage_attr}
\title{CDEC Mainstem Streamgage Attributes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 194 rows and 19 columns.
}
\usage{
streamgage_attr
}
\description{
CDEC Mainstem Streamgage Attributes
}
\keyword{datasets}
