% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamgage_da_attr}
\alias{streamgage_da_attr}
\title{CDEC Mainstem Streamgage Drainage Area Attributes}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 163 rows and 4 columns.
}
\usage{
streamgage_da_attr
}
\description{
CDEC Mainstem Streamgage Drainage Area Attributes
}
\keyword{datasets}
