% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamgage_duration_rating_curves}
\alias{streamgage_duration_rating_curves}
\title{CDEC Mainstem Streamgage Duration Rating Curves}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5230 rows and 5 columns.
}
\usage{
streamgage_duration_rating_curves
}
\description{
CDEC Mainstem Streamgage Duration Rating Curves
}
\keyword{datasets}
