% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamgage_geom}
\alias{streamgage_geom}
\title{CDEC Mainstem Streamgage Point Geometries}
\format{
An object of class \code{sf} (inherits from \code{grouped_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 194 rows and 2 columns.
}
\usage{
streamgage_geom
}
\description{
CDEC Mainstem Streamgage Point Geometries
}
\keyword{datasets}
