% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suitability_hsi_vel_spawning}
\alias{suitability_hsi_vel_spawning}
\title{Velocity suitability criteria for spawning by run}
\format{
An object of class \code{list} of length 5.
}
\usage{
suitability_hsi_vel_spawning
}
\description{
Velocity suitability criteria for spawning by run
}
\keyword{datasets}
