% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability.R
\name{suitability_postprocess}
\alias{suitability_postprocess}
\title{Postprocessor for both SRH-2D and HEC-RAS results, applying linear ft normalization by group}
\usage{
suitability_postprocess(result_tbl, group_polygons, .group_var = comid)
}
\arguments{
\item{result_tbl}{A \code{tbl_df} data frame with one row per group (\code{comid}) per flow (\code{flow_cfs}), containing the total inundated area (\code{area_tot}) in square feet, the suitable habitat area (\code{area_wua}) in square feet, and the ratio of suitable to total area (\code{area_pct}). This is the format returned by \code{raster_summarize_hsi()} and \code{vector_summarize_hsi()}.}

\item{group_polygons}{An \code{sf} polygon feature collection with one polygon for each analysis reach. These may have been delineated previously by splitting the hydraulic model domain into subsections.}

\item{.group_var}{The unquoted name of the attribute used to identify groups in the \code{group_polygons} layer. Defaults to \code{comid}}
}
\value{
A \code{tbl_df} data frame with one row per group (\code{comid}) per flow (\code{flow_cfs}), containing the total inundated area (\code{area_tot}) in square feet, the suitable habitat area (\code{area_wua}) in square feet, the ratio of suitable to total area (\code{area_pct}), the total inundated area per linear ft (\code{ind_per_lf}), and the suitable habitat area per linear ft (\code{wua_per_lf}).
}
\description{
Postprocessor for both SRH-2D and HEC-RAS results, applying linear ft normalization by group
}
