% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{trans_semiIHS}
\alias{trans_semiIHS}
\title{Semi Inverse Hyperbolic Sine Transform (scales transformation for ggplot)}
\format{
An object of class \code{trans} of length 7.
}
\usage{
trans_semiIHS
}
\description{
Semi Inverse Hyperbolic Sine Transform (scales transformation for ggplot)
}
\keyword{internal}
