% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_calculate_hsi}
\alias{vector_calculate_hsi}
\title{Calculate HSI for each row of SRH-2D results}
\usage{
vector_calculate_hsi(data, hsi_func = vector_dvhsi_hqt)
}
\arguments{
\item{data}{A \code{tbl_df} of the format returned by \code{vector_import_srh2d()}}

\item{hsi_func}{A function that takes a depth (ft) vector and velocity (ft/s) vector and returns a vector of habitat suitability index values between 0 and 1. Function arguments should be \code{d} and \code{v} in US units. Defaults to the \code{vector_dvhsi_hqt} function in this package.}
}
\value{
A \code{tbl_df} data frame containing the calculated habitat suitability indices by flow (\code{flow_cfs}) and vertex (\code{vid}) along with the associated depths and velocities.
}
\description{
Calculate HSI for each row of SRH-2D results
}
