% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_dvhsi_lyr}
\alias{vector_dvhsi_lyr}
\title{Vector Habitat Suitability Index (HSI) function: LYR}
\usage{
vector_dvhsi_lyr(d, v)
}
\arguments{
\item{d}{a vector of depths in feet}

\item{v}{a vector of velocities in feet per second}
}
\value{
A vector of habitat suitability index values between 0 and 1.
}
\description{
Vector Habitat Suitability Index (HSI) function: LYR
}
