% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_dvhsi_spawning}
\alias{vector_dvhsi_spawning}
\title{Vector Habitat Suitability Index (HSI) function: Fall-Run Chinook Spawning}
\usage{
vector_dvhsi_spawning(d, v, run = "max", strict = T)
}
\arguments{
\item{d}{a vector of depths in feet}

\item{v}{a vector of velocities in feet per second}

\item{run}{identifier for the salmon run of interest (one of \code{fall}, \code{late_fall}, \code{winter}, \code{spring}) or \code{max} to use the outer envelope of all runs. Defaults to \code{max}.}

\item{strict}{specify whether both criteria must strictly be met (minimum operation) or whether the factors are combined using a geometric mean}
}
\value{
A vector of habitat suitability index values between 0 and 1.
}
\description{
Vector Habitat Suitability Index (HSI) function: Fall-Run Chinook Spawning
}
