% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_prep_mesh}
\alias{vector_prep_mesh}
\title{Prep SRH-2D mesh, assigning groups and calculating areas}
\usage{
vector_prep_mesh(
  mesh,
  group_polygons = NULL,
  mask_polygons = NULL,
  .group_var = comid,
  .id_var = vid
)
}
\arguments{
\item{mesh}{An \code{sf} polygon feature collection with one polygon for each vertex (computation point) identified by a \code{vid} attribute. These may be Voronoi (aka Thiessen) polygons generated based on vertex points in a tool such as sf (\code{st_voronoi}), QGIS (Voronoi Polygons), or ArcGIS (Create Thiessen Polygons).}

\item{group_polygons}{An \code{sf} polygon feature collection with one polygon for each analysis reach. These may have been delineated previously by splitting the hydraulic model domain into subsections.}

\item{mask_polygons}{Optional: An \code{sf} polygon feature collection defining areas to be masked out of the HSI calculation. These areas will still be included in the total inundated area.}

\item{.group_var}{The unquoted name of the attribute used to identify groups in the \code{group_polygons} layer. Defaults to \code{comid}}

\item{.id_var}{The unquoted name of the attribute used to identify vertices in the \code{mesh} layer. Defaults to \code{vid}}
}
\value{
A \code{tbl_df} with one row per vertex (computation point), including the vertex id (e.g. \code{vid}), the associated group id (e.g. \code{comid}), and the area of the polygon in square feet.
}
\description{
Prep SRH-2D mesh, assigning groups and calculating areas
}
