% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_summarize_hsi}
\alias{vector_summarize_hsi}
\title{Summarize habitat suitability for SRH-2D results}
\usage{
vector_summarize_hsi(mesh_tbl, hsi_tbl, .group_var = comid, .id_var = vid)
}
\arguments{
\item{mesh_tbl}{A \code{tbl_df} with one row per vertex, containing group IDs and mesh cell areas, as created with \code{vector_prep_mesh()}}

\item{hsi_tbl}{A \code{tbl_df} with one row per vertex per flow, containing depths, velocities, and habitat suitability indices, as created with \code{vector_calculate_hsi()}}

\item{.group_var}{The unquoted name of the attribute used to identify groups in the \code{group_polygons} layer. Defaults to \code{comid}}

\item{.id_var}{The unquoted name of the attribute used to identify vertices in the \code{mesh} layer. Defaults to \code{vid}}
}
\value{
A \code{tbl_df} data frame with one row per group (\code{comid}) per flow (\code{flow_cfs}), containing the total inundated area (\code{area_tot}) in square feet, the suitable habitat area (\code{area_wua}) in square feet, and the ratio of suitable to total area (\code{area_pct}).
}
\description{
Summarize habitat suitability for SRH-2D results
}
