% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wua_hydraulic_interp}
\alias{wua_hydraulic_interp}
\alias{wua_hydraulic}
\title{WUA/LF by ComID by Flow (training dataset, interpolated)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11466 rows and 11 columns.
}
\usage{
wua_hydraulic
}
\description{
WUA/LF by ComID by Flow (training dataset, interpolated)
}
\examples{
wua_hydraulic_interp |> pillar::glimpse()
}
\keyword{datasets}
