% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wua_predicted}
\alias{wua_predicted}
\title{WUA/LF by ComID by Flow (predicted by statistical model)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1217164 rows and 10 columns.
}
\usage{
wua_predicted
}
\description{
WUA/LF by ComID by Flow (predicted by statistical model)
}
\examples{
wua_predicted |> pillar::glimpse()
}
\keyword{datasets}
