% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wua_hydraulic}
\alias{wua_hydraulic}
\title{WUA/LF by ComID by Flow (training dataset)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3822 rows and 9 columns.
}
\usage{
wua_hydraulic
}
\description{
WUA/LF by ComID by Flow (training dataset)
}
\examples{
wua_hydraulic |> pillar::glimpse()
}
\keyword{datasets}
